refactor: move operation cell to component package
