+++ packages/component/src/components/page-list/operation-menu-items/CopyLink.tsx
@@ -1,10 +1,8 @@
-import { MenuItem } from '@affine/component';
-import { useAFFiNEI18N } from '@affine/i18n/hooks';
+import { MenuItem, toast } from '@affine/component';
+import { useTranslation } from '@affine/i18n';
 import { CopyIcon } from '@blocksuite/icons';
 import { useCallback } from 'react';
 
-//
-import { toast } from '../../../utils';
 import type { CommonMenuItemProps } from './types';
 
 export const CopyLink = ({ onItemClick, onSelect }: CommonMenuItemProps) => {
