+++ packages/component/src/components/page-list/operation-menu-items/MoveToTrash.tsx
