+++ packages/component/src/components/page-list/operation-menu-items/index.ts
@@ -1,5 +1,5 @@
 export * from './CopyLink';
 export * from './DisablePublicSharing';
 export * from './Export';
-export * from './MoveTo';
+// export * from './MoveTo';
 export * from './MoveToTrash';
